+++ NEWS.md
@@ -1,5 +1,10 @@
 # 📰 News
 
+## v0.8.0
+
+- Add close_fd & check_unrar config options for API use.
+- Remove closefd constructor option.
+
 ## v0.7.1
 
 - Remove unrar.cffi support.